# Conflicts:
#	ExadelBonusPlus.Services.Models/MapperProfile.cs
#	ExadelBonusPlus.Services/Utils/EntityExtensions.cs

+++ ExadelBonusPlus.Services/VendorService.cs
@@ -1,4 +1,7 @@
-﻿using ExadelBonusPlus.Services.Models;
+﻿using AutoMapper;
+using ExadelBonusPlus.Services.Models;
+using ExadelBonusPlus.Services.Models.DTO;
+using ExadelBonusPlus.Services.Properties;
 using System;
 using System.Collections.Generic;
 using System.Threading;
@@ -9,28 +12,53 @@ namespace ExadelBonusPlus.Services
     public class VendorService : IVendorService
     {
         private IVendorRepository _vendorRepository;
-        public VendorService(IVendorRepository vendorRepository)
+        private readonly IMapper _mapper;
+
+        public VendorService(IVendorRepository vendorRepository, IMapper mapper)
         {
             _vendorRepository = vendorRepository;
+            _mapper = mapper;
         }
-        public Task AddVendorAsync(Vendor model, CancellationToken cancellationToken)
+        public async Task<VendorDto> AddVendorAsync(AddVendorDto model, CancellationToken cancellationToken)
         {
-            return _vendorRepository.AddAsync(model, cancellationToken);
+            if (model is null)
+            {
+                throw new ArgumentNullException("", Resources.ModelIsNull);
+            }
+            var vendor = _mapper.Map<Vendor>(model);
+            vendor.SetInitialValues(vendor);
+            await _vendorRepository.AddAsync(vendor, cancellationToken);
+            return _mapper.Map<VendorDto>(vendor);
+         }
+
+        public async Task<VendorDto> DeleteVendorAsync(Guid id, CancellationToken cancellationToken)
+        {
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentNullException("", Resources.IdentifierIsNull);
+            }
+
+            var result = await _vendorRepository.RemoveAsync(id, cancellationToken);
+
+            return _mapper.Map<VendorDto>(result);
         }
 
-        public Task DeleteVendorAsync(Guid id, CancellationToken cancellationToken)
+        public async Task<IEnumerable<VendorDto>> GetAllVendorsAsync(CancellationToken cancellationToken)
         {
-            return _vendorRepository.RemoveAsync(id, cancellationToken);
+            var result = await _vendorRepository.GetAllAsync(cancellationToken);
+
+            return _mapper.Map<IEnumerable<VendorDto>>(result);
         }
 
-        public Task<IEnumerable<Vendor>> GetAllVendorsAsync(CancellationToken cancellationToken)
+        public async Task<VendorDto> GetVendorByIdAsync(Guid id, CancellationToken cancellationToken)
         {
-            return _vendorRepository.GetAllAsync(cancellationToken);
-        }
+            if (id == Guid.Empty)
+            {
+                throw new ArgumentNullException("", Resources.IdentifierIsNull);
+            }
+            var result = await _vendorRepository.GetByIdAsync(id, cancellationToken);
+            return result is null ? throw new ArgumentException("", Resources.FindbyIdError) : _mapper.Map<VendorDto>(result);
 
-        public Task<Vendor> GetVendorByIdAsync(Guid id, CancellationToken cancellationToken)
-        {
-            return _vendorRepository.GetByIdAsync(id, cancellationToken);
         }
 
         public Task<IEnumerable<Vendor>> SearchVendorByLocationAsync(string city, CancellationToken cancellationToken)
@@ -38,9 +66,21 @@ namespace ExadelBonusPlus.Services
             return _vendorRepository.SearchVendorByLocationAsync(city, cancellationToken);
         }
 
-        public Task UpdateVendorAsync(Vendor model, CancellationToken cancellationToken)
+        public async Task<VendorDto> UpdateVendorAsync(Guid id, VendorDto model, CancellationToken cancellationToken)
         {
-            return _vendorRepository.UpdateAsync(model.Id, model, cancellationToken);
-        }
+            if(id == Guid.Empty)
+            {
+                throw new ArgumentNullException("", Resources.IdentifierIsNull);
+            }
+            if (model is null)
+            {
+                throw new ArgumentNullException("", Resources.ModelIsNull);
+            }
+            var vendor = _mapper.Map<Vendor>(model);
+
+            await _vendorRepository.UpdateAsync(id, vendor, cancellationToken);
+
+            return model;
+         }
     }
 }
